requirements.txt
# Parallel Image Processing System - Dependencies
#
# Install with: pip install -r requirements.txt

Pillow>=9.0.0
numpy>=1.21.0
